Updating timings
Here I updated some timings so the red monster and eye only appeared after a certain sound effect on the song. Since I was able to do this, I also changed the colour of the window initially so it looked like the monster wasn't there yet. And I made the glass cracks disappear when the monster appeared.