Added second monster
I managed to add the second monster in which is a mirrored copy of the original. I think this added symmetry in the design is more effective. Now I just need to work out how to get the spike on the end of the tail to move with the tail. Also need to tidy up the code and add more comments.